+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added "Force Despawn Bosses" to configuration, so that bosses spawned via effects can be forcibly despawned if all
   players are dead.
 - Added unique look-up identifier to "Features".
+- Added "Give Food" effect (disabled).
 - Changed description of "Enable Spawn Protection" configuration so that it is easier to understand.
 - Changed "Explode Inventory" drop chance additive value from 10 to 15 so that more items are dropped.
 - Changed "Rainbow Feet" effect to also cause killed NPCs to spawn multiple paint-ball projectiles.